Fix wrong data type of exp in runtime data

+++ RuntimeDataLib/Level.h
@@ -2,8 +2,8 @@ RUNTIME_DATA_FILE_BEGIN(level.enc)
 
 RUNTIME_DATA_TYPE_BEGIN(Level, "Level.level_up.con", 256)
 	RUNTIME_DATA_PROPERTY(Int32, Level, "level")
-	RUNTIME_DATA_PROPERTY(Int32, Exp, "exp")
-	RUNTIME_DATA_PROPERTY(Int32, AccumulatedExp, "accuexp")
+	RUNTIME_DATA_PROPERTY(UInt64, Exp, "exp")
+	RUNTIME_DATA_PROPERTY(UInt64, AccumulatedExp, "accuexp")
 RUNTIME_DATA_TYPE_END(Level)
 RUNTIME_DATA_TYPE_INDEX(Level, Int32, Level)
